long overdue refactoring of the codde evaluation module, split into multiple files
