+++ src/eval/index.ts
@@ -6,6 +6,10 @@ import { Parser } from "../lang/parser";
 import { createDefaultServices } from "../lang/interpreter/defaultService";
 import { Interpreter } from "../lang/interpreter/interpreter";
 import { escapeHtmlChars } from "../escape";
+import { isRecursive } from "./isRecursive";
+import { hasStructure } from "./hasStructure";
+//import { run_tscript } from "./run_tscript";
+import { run_multiple } from "./run_multiple";
 
 //
 // The functions in this module allow running and monitoring the runtime
@@ -514,877 +518,6 @@ export const evaluation = (function () {
 		}
 	}
 
-	// create an interpreter with specialized services for observing the
-	// runtime behavior of a program
-	function createInterpreter(program, inputs, output) {
-		let interpreter = new Interpreter(program, createDefaultServices());
-		interpreter.eventnames["canvas.resize"] = true;
-		interpreter.eventnames["canvas.mousedown"] = true;
-		interpreter.eventnames["canvas.mouseup"] = true;
-		interpreter.eventnames["canvas.mousemove"] = true;
-		interpreter.eventnames["canvas.mouseout"] = true;
-		interpreter.eventnames["canvas.keydown"] = true;
-		interpreter.eventnames["canvas.keyup"] = true;
-		interpreter.eventnames["timer"] = true;
-		interpreter.reset();
-		for (let i = 0; i < inputs.length; i++) {
-			if (typeof inputs[i] === "object" && inputs[i] !== null) {
-				let type = inputs[i].type;
-				let classname = inputs[i].classname;
-				let properties = inputs[i].event;
-
-				// name lookup of the event type (class)
-				let cls = interpreter.program;
-				let parts = classname.split(".");
-				for (let name of parts) cls = cls.names[name];
-
-				// create the event object
-				let event = { type: cls, value: { a: Array() } };
-				for (let i = 0; i < type.objectsize; i++) {
-					event.value.a.push(
-						TScript.json2typed.call(interpreter, null)
-					);
-				}
-
-				// fill in properties
-				if (properties) {
-					for (let key in properties) {
-						if (!properties.hasOwnProperty(key)) continue;
-						let index = cls.names[key].id;
-						event.value.a[index] = TScript.json2typed.call(
-							interpreter,
-							properties[key]
-						);
-					}
-				}
-
-				// enqueue event
-				interpreter.eventqueue.push({ type: type, event: event });
-
-				// remove from inputs
-				inputs.splice(i, 1);
-				i--;
-			}
-		}
-		let orig_turtle_move = interpreter.service.turtle.move;
-		interpreter.service.print = function (msg) {
-			output.push({ type: "print", value: msg });
-		};
-		interpreter.service.alert = function (msg) {
-			return new Promise((resolve, reject) => {
-				output.push({ type: "alert", value: msg });
-				resolve(null);
-			});
-		};
-		interpreter.service.confirm = function (msg) {
-			return new Promise((resolve, reject) => {
-				if (!inputs || inputs.length === 0) resolve(false);
-				else resolve(!!inputs.shift());
-			});
-		};
-		interpreter.service.prompt = function (msg) {
-			return new Promise((resolve, reject) => {
-				if (!inputs || inputs.length === 0) resolve("");
-				else resolve(inputs.shift());
-			});
-		};
-		interpreter.service.message = function (
-			msg,
-			line: any = null,
-			ch: any = null,
-			href: any = ""
-		) {
-			output.push({
-				type: "runtime error",
-				message: msg,
-				line: line,
-				href: href,
-			});
-		};
-		interpreter.service.turtle.move = function (distance) {
-			let x0 = interpreter.service.turtle.x;
-			let y0 = interpreter.service.turtle.y;
-			orig_turtle_move.call(this, distance);
-			let x1 = interpreter.service.turtle.x;
-			let y1 = interpreter.service.turtle.y;
-			if (interpreter.service.turtle.down && distance != 0)
-				output.push({
-					type: "turtle line",
-					from: [x0, y0],
-					to: [x1, y1],
-					color: interpreter.service.turtle.rgb,
-				});
-		};
-		interpreter.service.canvas.dom = { width: 1920, height: 1080 };
-		interpreter.service.canvas.setLineColor = function (r, g, b) {
-			output.push({
-				type: "canvas setLineColor",
-				r: r,
-				g: g,
-				b: b,
-			});
-		};
-		interpreter.service.canvas.setFillColor = function (r, g, b) {
-			output.push({
-				type: "canvas setFillColor",
-				r: r,
-				g: g,
-				b: b,
-			});
-		};
-		interpreter.service.canvas.clear = function () {
-			output.push({ type: "canvas clear" });
-		};
-		interpreter.service.canvas.line = function (x1, y1, x2, y2) {
-			output.push({
-				type: "canvas line",
-				x1: x1,
-				y1: y1,
-				x2: x2,
-				y2: y2,
-			});
-		};
-		interpreter.service.canvas.rect = function (left, top, width, height) {
-			output.push({
-				type: "canvas rect",
-				left: left,
-				top: top,
-				width: width,
-				height: height,
-			});
-		};
-		interpreter.service.canvas.fillRect = function (
-			left,
-			top,
-			width,
-			height
-		) {
-			output.push({
-				type: "canvas fillRect",
-				left: left,
-				top: top,
-				width: width,
-				height: height,
-			});
-		};
-		interpreter.service.canvas.frameRect = function (
-			left,
-			top,
-			width,
-			height
-		) {
-			output.push({
-				type: "canvas frameRect",
-				left: left,
-				top: top,
-				width: width,
-				height: height,
-			});
-		};
-		interpreter.service.canvas.circle = function (x, y, radius) {
-			output.push({ type: "canvas circle", x: x, y: y, radius: radius });
-		};
-		interpreter.service.canvas.fillCircle = function (x, y, radius) {
-			output.push({
-				type: "canvas fillCircle",
-				x: x,
-				y: y,
-				radius: radius,
-			});
-		};
-		interpreter.service.canvas.frameCircle = function (x, y, radius) {
-			output.push({
-				type: "canvas frameCircle",
-				x: x,
-				y: y,
-				radius: radius,
-			});
-		};
-		interpreter.service.canvas.curve = function (points, closed) {
-			output.push({
-				type: "canvas curve",
-				points: points,
-				closed: closed,
-			});
-		};
-		interpreter.service.canvas.fillArea = function (points) {
-			output.push({ type: "canvas fillArea", points: points });
-		};
-		interpreter.service.canvas.frameArea = function (points) {
-			output.push({ type: "canvas frameArea", points: points });
-		};
-		interpreter.service.canvas.text = function (x, y, str) {
-			output.push({ type: "canvas text", x: x, y: y, str: str });
-		};
-		interpreter.service.canvas.reset = function () {
-			output.push({ type: "canvas reset" });
-		};
-		interpreter.service.canvas.shift = function (dx, dy) {
-			output.push({ type: "canvas shift", dx: dx, dy: dy });
-		};
-		interpreter.service.canvas.scale = function (factor) {
-			output.push({ type: "canvas scale", factor: factor });
-		};
-		interpreter.service.canvas.rotate = function (angle) {
-			output.push({ type: "canvas rotate", angle: angle });
-		};
-		interpreter.service.canvas.transform = function (A, b) {
-			output.push({ type: "canvas transform", A: A, b: b });
-		};
-
-		return interpreter;
-	}
-
-	// This function runs a TScript program. It returns an array of
-	// events, like print messages, turtle and canvas outputs.
-	// Parameters:
-	//  - code is the TScript source code to run
-	//  - maxseconds is the timeout, default=3
-	//  - inputs in an array of values returned by consecutive calls to TScript's confirm or prompt, default=[]
-	module.run_tscript = function (code, maxseconds = 3.0, inputs = []) {
-		inputs = inputs.slice();
-		let output = new Array();
-
-		let result = Parser.parse(code);
-		let errors = result.errors;
-		if (errors.length > 0) {
-			for (let i = 0; i < errors.length; i++) {
-				let err = errors[i];
-				output.push({
-					type: "compile error",
-					line: err.line,
-					message: err.message,
-				});
-			}
-			return output;
-		}
-		let program = result.program;
-
-		// create an interpreter and prepare it for collecting output events
-		let interpreter = createInterpreter(program, inputs, output);
-		interpreter.stopthread();
-		interpreter.reset();
-
-		// run the program
-		let timeout = new Date().getTime() + 1000 * maxseconds;
-		while (true) {
-			let start = new Date().getTime();
-			if (start >= timeout) {
-				output.push({
-					type: "timeout - program execution took too long",
-				});
-				return output;
-			}
-
-			if (interpreter.status === "waiting")
-				interpreter.status = "running";
-			if (interpreter.status != "running") break;
-
-			while (
-				new Date().getTime() - start < 14 &&
-				interpreter.status === "running"
-			) {
-				interpreter.exec_step();
-				if (interpreter.halt) break;
-			}
-			if (!interpreter.timerEventEnqueued) {
-				interpreter.timerEventEnqueued = interpreter.enqueueEvent(
-					"timer",
-					{
-						type: interpreter.program.types[Typeid.typeid_null],
-						value: { b: null },
-					}
-				);
-			}
-		}
-
-		return output;
-	};
-
-	// This function runs multiple TScript programs. It returns immediately.
-	// The programs are run asynchronously in the background until they finish or time out.
-	// Parameters:
-	//  - code is an array of TScript programs or JS test functions
-	//  - inputs is an array of arrays, containing values returned by consecutive calls to TScript's confirm or prompt.
-	//  - maxseconds is the timeout
-	//  - process is a function taking the array of event arrays as its argument for further processing.
-	module.run_multiple = function (code, inputs, maxseconds, process) {
-		let timeout = new Date().getTime() + 1000 * maxseconds;
-
-		let index = 0;
-		let all: any = new Array();
-		let output = new Array();
-		for (let i = 0; i < inputs.length; i++) inputs[i] = inputs[i].slice();
-
-		function compute() {
-			let c = code[index];
-			if (c[0] === "@") {
-				let split = c.indexOf("\n@\n");
-				let jsc = c.substr(1, split - 1);
-				let tsc = c.substr(split + 3);
-				let testfunction;
-				eval(
-					"testfunction = function(code, parse, hasStructure, isRecursive) {\n" +
-						jsc +
-						"\nreturn null; };\n"
-				);
-				all.push(
-					testfunction(
-						tsc,
-						Parser.parse,
-						module.hasStructure,
-						module.isRecursive
-					)
-				);
-				output = new Array();
-				index++;
-				if (index === code.length) process(all);
-				else compute();
-			} else {
-				let result = Parser.parse(c);
-				let errors = result.errors;
-				if (errors.length > 0) {
-					for (let i = 0; i < errors.length; i++) {
-						let err = errors[i];
-						output.push({
-							type: "compile error",
-							line: err.line,
-							message: err.message,
-						});
-					}
-
-					// output the collected output
-					all.push(output);
-					output = new Array();
-					index++;
-					if (index === code.length) process(all);
-					else compute();
-				} else {
-					let program = result.program;
-
-					// create an interpreter and prepare it for collecting output events
-					let interpreter = createInterpreter(
-						program,
-						inputs[index],
-						output
-					);
-					interpreter.reset();
-
-					// run the program
-					interpreter.run();
-
-					// stop when the program has finished
-					function monitor() {
-						if (new Date().getTime() > timeout) {
-							interpreter.stopthread();
-							all = "timeout - program execution took too long";
-							process(all);
-						} else if (
-							interpreter.status === "finished" ||
-							interpreter.status === "error"
-						) {
-							interpreter.stopthread();
-
-							// output the collected output
-							all.push(output);
-							output = new Array();
-							index++;
-							if (index === code.length) process(all);
-							else compute();
-						} else setTimeout(monitor, 1);
-					}
-					setTimeout(monitor, 1);
-				}
-			}
-		}
-		compute();
-	};
-
-	const reserved_node_names = {
-		breakpoints: true,
-		children: true,
-		class_constructor: true,
-		declaration: true,
-		displayname: true,
-		id: true,
-		lines: true,
-		name: true,
-		names: true,
-		operator: true,
-		params: true,
-		parent: true,
-		petype: true,
-		reference: true,
-		scope: true,
-		step: true,
-		sim: true,
-		superclass: true,
-		type: true,
-		types: true,
-		typedvalue: true,
-		var_id: true,
-		var_scope: true,
-		variables: true,
-		where: true,
-	};
-
-	// static code analysis, try to decide whether a program contains a recursive function or not
-	function isRecursiveStatic(program) {
-		if (!program) return false;
-
-		function isObject(value) {
-			return (
-				value !== null &&
-				typeof value === "object" &&
-				value.constructor === Object
-			);
-		}
-
-		function isNode(value) {
-			return Array.isArray(value) || isObject(value);
-		}
-
-		function rec(node, fnames) {
-			if (node.builtin) return false;
-			if (node.petype === "breakpoint") return false;
-
-			if (node.petype === "function" && node.name) {
-				let names = fnames.slice();
-				names.push(node.name);
-				fnames = names;
-			}
-			if (node.petype === "function call") {
-				if (node.base.name && fnames.indexOf(node.base.name) >= 0)
-					return true;
-				if (node.base.petype === "this") return true;
-			}
-			for (let key in node) {
-				if (!node.hasOwnProperty(key)) continue;
-				if (reserved_node_names.hasOwnProperty(key)) continue;
-				let n = node[key];
-				if (isNode(n)) {
-					if (rec(n, fnames)) return true;
-				}
-			}
-			return false;
-		}
-
-		return rec(program, new Array());
-	}
-
-	// dynamic behavior analysis, try to decide whether a program contains a recursive function or not
-	function isRecursiveDynamic(program, maxseconds = 3.0, inputs = []) {
-		if (!program) return false;
-		inputs = inputs.slice();
-
-		// create an interpreter with empty services
-		let interpreter = new Interpreter(program, createDefaultServices());
-		let orig_turtle_move = interpreter.service.turtle.move;
-		interpreter.service.print = function (msg) {};
-		interpreter.service.alert = function (msg) {};
-		interpreter.service.message = function (msg, line, ch, href) {};
-		interpreter.service.prompt = function (msg) {};
-		interpreter.service.turtle.move = function (distance) {};
-		interpreter.service.canvas.dom = { width: 1920, height: 1080 };
-		interpreter.service.canvas.clear = function () {};
-		interpreter.service.canvas.line = function (x1, y1, x2, y2) {};
-		interpreter.service.canvas.rect = function (
-			left,
-			top,
-			width,
-			height
-		) {};
-		interpreter.service.canvas.fillRect = function (
-			left,
-			top,
-			width,
-			height
-		) {};
-		interpreter.service.canvas.frameRect = function (
-			left,
-			top,
-			width,
-			height
-		) {};
-		interpreter.service.canvas.circle = function (x, y, radius) {};
-		interpreter.service.canvas.fillCircle = function (x, y, radius) {};
-		interpreter.service.canvas.frameCircle = function (x, y, radius) {};
-		interpreter.service.canvas.curve = function (points, closed) {};
-		interpreter.service.canvas.fillArea = function (points) {};
-		interpreter.service.canvas.frameArea = function (points) {};
-		interpreter.service.canvas.text = function (x, y, str) {};
-		interpreter.service.canvas.reset = function () {};
-		interpreter.service.canvas.shift = function (dx, dy) {};
-		interpreter.service.canvas.scale = function (factor) {};
-		interpreter.service.canvas.rotate = function (angle) {};
-		interpreter.service.canvas.transform = function (A, b) {};
-		interpreter.stopthread();
-		interpreter.reset();
-
-		// run the program
-		let timeout = new Date().getTime() + 1000 * maxseconds;
-		while (true) {
-			let start = new Date().getTime();
-			if (start >= timeout) break; // not ideal, but better than an infinite loop
-
-			if (interpreter.status === "waiting")
-				interpreter.status = "running";
-			if (interpreter.status != "running") break;
-
-			while (
-				new Date().getTime() - start < 14 &&
-				interpreter.status === "running"
-			) {
-				interpreter.exec_step();
-
-				// check for recursion
-				let n = interpreter.stack.length;
-				for (let i = 0; i < n; i++) {
-					for (let j = 0; j < i; j++) {
-						if (
-							interpreter.stack[i].pe[0] ===
-							interpreter.stack[j].pe[0]
-						)
-							return true;
-					}
-				}
-
-				if (interpreter.halt) break;
-			}
-			if (!interpreter.timerEventEnqueued) {
-				interpreter.timerEventEnqueued = interpreter.enqueueEvent(
-					"timer",
-					{
-						type: interpreter.program.types[Typeid.typeid_null],
-						value: { b: null },
-					}
-				);
-			}
-		}
-
-		return false;
-	}
-
-	// Return true if the given (translated) program contains at least one
-	// recursive function. The test is performed statically at first. If no
-	// recursion is found then the program is executed -- hence make sure to
-	// include test code that actually invokes the recursion -- and the
-	// runtime behavior is analyzed.
-	module.isRecursive = function (program) {
-		if (isRecursiveStatic(program)) return true;
-		else return isRecursiveDynamic(program);
-	};
-
-	// Return true if the given program (more precisely, its AST)
-	// has a structure that is compatible with the given pseudo code.
-	module.hasStructure = function (program, pseudo) {
-		function compilePseudo(code) {
-			let pos = 0;
-
-			function skipwhite() {
-				while (
-					pos < code.length &&
-					(code[pos] === " " ||
-						code[pos] === "\t" ||
-						code[pos] === "\n" ||
-						code[pos] === "\r")
-				)
-					pos++;
-			}
-
-			function identifier() {
-				skipwhite();
-				if (pos >= code.length) return null;
-				let c = code[pos];
-				if (c === '"') {
-					let ret = "";
-					pos++;
-					while (pos < code.length) {
-						c = code[pos];
-						pos++;
-						if (c === '"') return ret;
-						else ret += c;
-					}
-					return null;
-				} else if (c === "'") {
-					let ret = "";
-					pos++;
-					while (pos < code.length) {
-						c = code[pos];
-						pos++;
-						if (c === "'") return ret;
-						else ret += c;
-					}
-					return null;
-				} else if (
-					(c >= "A" && c <= "Z") ||
-					(c >= "a" && c <= "z") ||
-					c === "_"
-				) {
-					let ret = c;
-					pos++;
-					while (true) {
-						c = code[pos];
-						if (
-							(c >= "A" && c <= "Z") ||
-							(c >= "a" && c <= "z") ||
-							(c >= "0" && c <= "9") ||
-							c === "_"
-						) {
-							ret += c;
-							pos++;
-						} else break;
-					}
-					return ret;
-				} else return null;
-			}
-
-			function command() {
-				let ident = identifier();
-				if (ident === null) return null;
-
-				let ret: any = { petype: ident };
-
-				skipwhite();
-				if (pos >= code.length) return null;
-				let c = code[pos];
-				if (c === "(") {
-					pos++;
-					let name = identifier();
-					if (name === null) return null;
-					skipwhite();
-					let c = code[pos];
-					if (c != ")") return null;
-					pos++;
-					ret.base = { name: name };
-				} else if (c === "[") {
-					pos++;
-					let name = identifier();
-					if (name === null) return null;
-					skipwhite();
-					let c = code[pos];
-					if (c != "]") return null;
-					pos++;
-					ret.name = name;
-				}
-
-				skipwhite();
-				if (pos >= code.length) return null;
-				c = code[pos];
-				if (c === "{") {
-					pos++;
-					let sub = new Array();
-					while (true) {
-						skipwhite();
-						if (pos >= code.length) return null;
-						if (code[pos] === ";") {
-							pos++;
-							continue;
-						}
-						if (code[pos] === "}") {
-							pos++;
-							ret.sub = sub.length === 1 ? sub[0] : sub;
-							return ret;
-						}
-						let cmd = command();
-						if (cmd === null) return null;
-						sub.push(cmd);
-					}
-				} else if (c === ";") {
-					pos++;
-					return ret;
-				} else return null;
-			}
-
-			skipwhite();
-			if (pos >= code.length) return null;
-			let c = code[pos];
-			if (c === "{") {
-				pos++;
-				let ret = new Array();
-				while (true) {
-					skipwhite();
-					if (pos >= code.length) return null;
-					if (code[pos] === ";") {
-						pos++;
-						continue;
-					}
-					if (code[pos] === "}") {
-						pos++;
-						skipwhite();
-						if (pos >= code.length)
-							return ret.length === 1 ? ret[0] : ret;
-						else return null;
-					}
-					let cmd = command();
-					if (cmd === null) return null;
-					ret.push(cmd);
-				}
-			} else {
-				let cmd = command();
-				skipwhite();
-				if (pos >= code.length) return cmd;
-				else return null;
-			}
-		}
-
-		// The function returns true if key is a sub-tree of data, and false otherwise.
-		// A tree is an arbitrary array or object. Non-array non-object nodes are ignored.
-		// Many member names like "parent" and "petype" are reserved. All other object members are considered child nodes.
-		// Two nodes match under the following conditions:
-		// - Arrays can match arrays; objects can match objects.
-		// - If at least one of the nodes has a petype member then both need one, and the key petype must be a substring of the data petype.
-		// - If at least one of the nodes has a name member then both need one, and the key name must match the data name.
-		// - Array members are considered in their natural order.
-		// - In the key tree, objects must have at most one child node, hence there is no order ambiguity.
-		function isSubtree(key, data) {
-			function isObject(value) {
-				return (
-					value !== null &&
-					typeof value === "object" &&
-					value.constructor === Object
-				);
-			}
-
-			function isNode(value) {
-				return Array.isArray(value) || isObject(value);
-			}
-
-			function equal(c1, c2) {
-				if (
-					c1.nstack.length != c2.nstack.length ||
-					c1.dstack.length != c2.dstack.length ||
-					c1.next.length != c2.next.length
-				)
-					return false;
-				for (let i = 0; i < c1.nstack.length; i++) {
-					if (c1.nstack[i].length != c2.nstack[i].length)
-						return false;
-					for (let j = 0; j < c1.nstack[i].length; j++) {
-						if (c1.nstack[i][j] !== c2.nstack[i][j]) return false;
-					}
-				}
-				for (let i = 0; i < c1.dstack.length; i++) {
-					if (c1.dstack[i] !== c2.dstack[i]) return false;
-				}
-				for (let i = 0; i < c1.next.length; i++) {
-					if (c1.next[i] !== c2.next[i]) return false;
-				}
-				return true;
-			}
-
-			// initial cursor
-			let cs = [
-				{ nstack: new Array(), dstack: new Array(), next: [key] },
-			];
-
-			// recursive traversal of the data tree
-			function rec(node) {
-				if (!node) return false;
-				if (node.builtin) return false;
-				if (node.petype === "breakpoint") return false;
-
-				// forward pass: match nodes
-				let ii = cs.length;
-				for (let i = 0; i < ii; i++) {
-					let c = cs[i];
-					if (c.next.length === 0) continue; // this cursor is waiting for an ancestor of the current data node
-					let next = c.next[0];
-					if (!isNode(next))
-						throw new Error(
-							"[isSubtree] key node is not an array or object"
-						);
-
-					// do the nodes match?
-					let match = Array.isArray(node) === Array.isArray(next);
-					if (match && isObject(next)) {
-						if (next.hasOwnProperty("petype")) {
-							let d = node.petype ? node.petype : "";
-							let k = next.petype ? next.petype : "";
-							match = d.indexOf(k) >= 0;
-						}
-						if (next.hasOwnProperty("name")) {
-							let d = node.name ? node.name : "";
-							let k = next.name ? next.name : "";
-							match = d === k;
-						}
-					}
-
-					if (match) {
-						// match found: create a new cursor
-						// - create a new next array from the children of the key node
-						// - add old next array with the first position removed and the data node to the stacks
-						let ns = c.nstack.slice();
-						let ds = c.dstack.slice();
-						ns.push(c.next.slice(1));
-						ds.push(node);
-						let nx = new Array();
-						for (let key in next) {
-							if (!next.hasOwnProperty(key)) continue;
-							if (reserved_node_names.hasOwnProperty(key))
-								continue;
-							let n = next[key];
-							if (isNode(n)) nx.push(n);
-						}
-						cs.push({ nstack: ns, dstack: ds, next: nx });
-					}
-				}
-
-				// recursion
-				for (let key in node) {
-					if (!node.hasOwnProperty(key)) continue;
-					if (reserved_node_names.hasOwnProperty(key)) continue;
-					let n = node[key];
-					if (isNode(n)) {
-						if (rec(n)) return true;
-					}
-				}
-
-				// backward pass: collect results
-				let cs1 = new Array(),
-					cs2 = new Array();
-				for (let i = 0; i < cs.length; i++) {
-					let c = cs[i];
-					if (c.dstack[c.dstack.length - 1] === node) {
-						if (c.next.length > 0) continue; // sub-tree match not completed: drop the cursor
-
-						// move the cursor to the next sub-tree
-						c.dstack.pop();
-						c.next = c.nstack.pop();
-						if (c.nstack.length === 0) return true; // success!
-						cs2.push(c);
-					} else cs1.push(c);
-				}
-
-				// check for duplicate cursors
-				cs = cs1;
-				for (let i = 0; i < cs2.length; i++) {
-					let duplicate = false;
-					for (let j = 0; j < cs.length; j++) {
-						if (equal(cs2[i], cs[j])) {
-							duplicate = true;
-							break;
-						}
-					}
-					if (!duplicate) cs.push(cs2[i]);
-				}
-
-				return false;
-			}
-
-			return rec(data);
-		}
-
-		let pc = compilePseudo(pseudo);
-		//		console.log("[compilePseudo]", pseudo, pc, program.commands.slice(34));
-		if (pc === null) {
-			console.log("error in pseudo code:\n\n" + pseudo);
-			return false;
-		}
-		// console.log(pc, program);
-		return isSubtree(pc, program);
-	};
-
 	// Evaluate a submission (TScript code) for a programming task. The
 	// task is specified as an object with the following keys:
 	//  - correct: This is the reference solution as a string.
@@ -1499,7 +632,7 @@ export const evaluation = (function () {
 		}
 
 		// run the code
-		module.run_multiple(codes, inputs, timeout, function (result) {
+		run_multiple(codes, inputs, timeout, function (result) {
 			let error: null | string = null;
 			let details: null | string = null;
 			let points = total;
