+++ src/eval/reserved_node_names.ts
@@ -1,0 +1,27 @@
+export const reserved_node_names = {
+	breakpoints: true,
+	children: true,
+	class_constructor: true,
+	declaration: true,
+	displayname: true,
+	id: true,
+	lines: true,
+	name: true,
+	names: true,
+	operator: true,
+	params: true,
+	parent: true,
+	petype: true,
+	reference: true,
+	scope: true,
+	step: true,
+	sim: true,
+	superclass: true,
+	type: true,
+	types: true,
+	typedvalue: true,
+	var_id: true,
+	var_scope: true,
+	variables: true,
+	where: true,
+};
